Feat: 동기로 분리
