--- src/main/java/stdev/domain/dream/presentation/dto/request/ContentGenerationRequest.java
@@ -1,9 +1,0 @@
-package stdev.domain.dream.presentation.dto.request;
-
-import lombok.Data;
-
-public record ContentGenerationRequest(
-        String topic
-) {
-
-}
