+++ src/main/java/stdev/domain/dreamanalysis/application/service/DreamServiceImpl.java
@@ -1,14 +1,13 @@
-package stdev.domain.dream.application.service;
+package stdev.domain.dreamanalysis.application.service;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import stdev.domain.dream.application.DreamService;
-import stdev.domain.dream.domain.entity.Dream;
-import stdev.domain.dream.domain.repository.DreamRepository;
-import stdev.domain.dream.presentation.dto.request.DreamRecordRequest;
+import stdev.domain.dreamanalysis.application.DreamService;
+import stdev.domain.dreamanalysis.domain.entity.Dream;
+import stdev.domain.dreamanalysis.domain.repository.DreamRepository;
 import stdev.global.config.FileStore;
 
 import java.io.IOException;
