+++ src/main/java/stdev/domain/dreamanalysis/domain/repository/DreamRepository.java
@@ -1,7 +1,7 @@
-package stdev.domain.dream.domain.repository;
+package stdev.domain.dreamanalysis.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import stdev.domain.dream.domain.entity.Dream;
+import stdev.domain.dreamanalysis.domain.entity.Dream;
 
 public interface DreamRepository extends JpaRepository<Dream, Long> {
 }
