+++ src/main/java/stdev/domain/dreamanalysis/presentation/dto/request/ContentGenerationRequest.java
@@ -1,0 +1,7 @@
+package stdev.domain.dreamanalysis.presentation.dto.request;
+
+public record ContentGenerationRequest(
+        String topic
+) {
+
+}
