+++ src/main/java/stdev/domain/dreamanalysis/presentation/dto/response/DreamCommentResponse.java
@@ -1,0 +1,11 @@
+package stdev.domain.dreamanalysis.presentation.dto.response;
+
+public record DreamCommentResponse(
+        String comment,
+
+        String category
+) {
+    public static DreamCommentResponse of(String comment, String category){
+        return new DreamCommentResponse(comment,category);
+    }
+}
