+++ src/main/java/stdev/domain/dreamanalysis/presentation/dto/response/DreamImageResponse.java
@@ -1,0 +1,9 @@
+package stdev.domain.dreamanalysis.presentation.dto.response;
+
+public record DreamImageResponse(
+        String imageUrl
+) {
+    public static DreamImageResponse of(String imageUrl) {
+        return new DreamImageResponse(imageUrl);
+    }
+}
